requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
# dlib
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.2
flatbuffers==22.10.26
fonttools==4.38.0
humanfriendly==10.0
idna==3.4
imageio==2.22.2
importlib-metadata==5.0.0
importlib-resources==5.10.1
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isort==5.10.1
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.4
jupyter_client==7.4.8
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.2.1
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
onnxruntime-gpu==1.13.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.9
psutil==5.9.4
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.5
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
scikit-image==0.17.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
# soupsieve==2.3.2.post1
sympy==1.10.1
terminado==0.17.1
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
# torch==1.12.0+cu113
torchstat==0.0.7
torchsummary==1.5.1
# torchvision==0.13.0+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
zipp==3.10.0
nibabel==4.0.2
